modularización y refactorizaciónde CreatePro duct
