+++ client/src/redux/actions/products.js
@@ -57,7 +57,6 @@ export const getProductDetail = (id) => {
 export const createProduct = (payload) => {
   return async (dispatch) => {
     try {
-      console.log("payload:",JSON.stringify(payload))
       let response = await axios.post(
         "http://localhost:3001/products/create",
         payload
